iris: refactor serial driver to use new IRQ mechanism
