+++ iris/arch/x86/core/print.cpp
@@ -8,13 +8,6 @@ void log_output_character(c32 value) {
         log_output_byte(di::Byte(byte));
     }
 }
-
-void log_output_byte(di::Byte byte) {
-    while ((x86::amd64::io_in<di::Byte>(0x3F8 + 5) & 0x20_b) == 0_b)
-        ;
-
-    x86::amd64::io_out(0x3F8, byte);
-}
 }
 
 namespace di::assert::detail {
