+++ iris/include/iris/arch/x86/amd64/hw/serial.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace iris::x86::amd64 {
+void init_serial_early_boot();
+void init_serial();
+}
